Fix issue in ChDeformableTire with setting coefficients of friction and restitution.
Previously, changing these values was possible only through the function SetContactMaterialProperties; i.e., at the same time as Y and nu.   Now, these two coefficients are set through separate functions, regardless of whether material properties or user-provided coefficients are specified.

Note that this involved a change in the format of the JSON files for ANCF tires and FEA tires.